Move presence class under API to avoid import issues.
