+++ synapse/api/presence.py
